@@ -1,27 +1,27 @@
 {
-  "_from": "ansi-regex@^3.0.0",
-  "_id": "ansi-regex@3.0.0",
+  "_from": "ansi-regex@^4.1.0",
+  "_id": "ansi-regex@4.1.0",
   "_inBundle": false,
-  "_integrity": "sha1-7QMXwyIGT3lGbAKWa922Bas32Zg=",
+  "_integrity": "sha512-1apePfXM1UOSqw0o9IiFAovVz9M5S1Dg+4TrDwfMewQ6p/rmMueb7tWZjQ1rx4Loy1ArBggoqGpfqqdI4rondg==",
   "_location": "/ansi-regex",
   "_phantomChildren": {},
   "_requested": {
     "type": "range",
     "registry": true,
-    "raw": "ansi-regex@^3.0.0",
+    "raw": "ansi-regex@^4.1.0",
     "name": "ansi-regex",
     "escapedName": "ansi-regex",
-    "rawSpec": "^3.0.0",
+    "rawSpec": "^4.1.0",
     "saveSpec": null,
-    "fetchSpec": "^3.0.0"
+    "fetchSpec": "^4.1.0"
   },
   "_requiredBy": [
     "/strip-ansi"
   ],
-  "_resolved": "https://registry.npmjs.org/ansi-regex/-/ansi-regex-3.0.0.tgz",
-  "_shasum": "ed0317c322064f79466c02966bddb605ab37d998",
-  "_spec": "ansi-regex@^3.0.0",
-  "_where": "C:\\Users\\ahmad\\Desktop\\2dv513\\node_modules\\strip-ansi",
+  "_resolved": "https://registry.npmjs.org/ansi-regex/-/ansi-regex-4.1.0.tgz",
+  "_shasum": "8b9f8f08cf1acb843756a839ca8c7e3168c51997",
+  "_spec": "ansi-regex@^4.1.0",
+  "_where": "C:\\Users\\ahmad\\Desktop\\2dv513-Assignment-3-Database-Theory\\backend_nodeJs\\node_modules\\strip-ansi",
   "author": {
     "name": "Sindre Sorhus",
     "email": "sindresorhus@gmail.com",
@@ -34,11 +34,11 @@
   "deprecated": false,
   "description": "Regular expression for matching ANSI escape codes",
   "devDependencies": {
-    "ava": "*",
-    "xo": "*"
+    "ava": "^0.25.0",
+    "xo": "^0.23.0"
   },
   "engines": {
-    "node": ">=4"
+    "node": ">=6"
   },
   "files": [
     "index.js"
@@ -81,5 +81,5 @@
     "test": "xo && ava",
     "view-supported": "node fixtures/view-codes.js"
   },
-  "version": "3.0.0"
+  "version": "4.1.0"
 }
