@@ -1,56 +1,56 @@
 {
-  "_from": "ansi-styles@^3.2.1",
-  "_id": "ansi-styles@3.2.1",
+  "_from": "ansi-styles@^4.1.0",
+  "_id": "ansi-styles@4.3.0",
   "_inBundle": false,
-  "_integrity": "sha512-VT0ZI6kZRdTh8YyJw3SMbYm/u+NqfsAxEpWO0Pf9sq8/e94WxxOpPKx9FR1FlyCtOVDNOQ+8ntlqFxiRc+r5qA==",
+  "_integrity": "sha512-zbB9rCJAT1rbjiVDb2hqKFHNYLxgtk8NURxZ3IZwD3F6NtxbXZQCnnSi1Lkx+IDohdPlFp222wVALIheZJQSEg==",
   "_location": "/ansi-styles",
   "_phantomChildren": {},
   "_requested": {
     "type": "range",
     "registry": true,
-    "raw": "ansi-styles@^3.2.1",
+    "raw": "ansi-styles@^4.1.0",
     "name": "ansi-styles",
     "escapedName": "ansi-styles",
-    "rawSpec": "^3.2.1",
+    "rawSpec": "^4.1.0",
     "saveSpec": null,
-    "fetchSpec": "^3.2.1"
+    "fetchSpec": "^4.1.0"
   },
   "_requiredBy": [
     "/chalk"
   ],
-  "_resolved": "https://registry.npmjs.org/ansi-styles/-/ansi-styles-3.2.1.tgz",
-  "_shasum": "41fbb20243e50b12be0f04b8dedbf07520ce841d",
-  "_spec": "ansi-styles@^3.2.1",
-  "_where": "C:\\Users\\ahmad\\Desktop\\2dv513\\node_modules\\chalk",
+  "_resolved": "https://registry.npmjs.org/ansi-styles/-/ansi-styles-4.3.0.tgz",
+  "_shasum": "edd803628ae71c04c85ae7a0906edad34b648937",
+  "_spec": "ansi-styles@^4.1.0",
+  "_where": "C:\\Users\\ahmad\\Desktop\\2dv513-Assignment-3-Database-Theory\\backend_nodeJs\\node_modules\\chalk",
   "author": {
     "name": "Sindre Sorhus",
     "email": "sindresorhus@gmail.com",
     "url": "sindresorhus.com"
-  },
-  "ava": {
-    "require": "babel-polyfill"
   },
   "bugs": {
     "url": "https://github.com/chalk/ansi-styles/issues"
   },
   "bundleDependencies": false,
   "dependencies": {
-    "color-convert": "^1.9.0"
+    "color-convert": "^2.0.1"
   },
   "deprecated": false,
   "description": "ANSI escape codes for styling strings in the terminal",
   "devDependencies": {
-    "ava": "*",
-    "babel-polyfill": "^6.23.0",
+    "@types/color-convert": "^1.9.0",
+    "ava": "^2.3.0",
     "svg-term-cli": "^2.1.1",
-    "xo": "*"
+    "tsd": "^0.11.0",
+    "xo": "^0.25.3"
   },
   "engines": {
-    "node": ">=4"
+    "node": ">=8"
   },
   "files": [
-    "index.js"
+    "index.js",
+    "index.d.ts"
   ],
+  "funding": "https://github.com/chalk/ansi-styles?sponsor=1",
   "homepage": "https://github.com/chalk/ansi-styles#readme",
   "keywords": [
     "ansi",
@@ -82,7 +82,7 @@
   },
   "scripts": {
     "screenshot": "svg-term --command='node screenshot' --out=screenshot.svg --padding=3 --width=55 --height=3 --at=1000 --no-cursor",
-    "test": "xo && ava"
+    "test": "xo && ava && tsd"
   },
-  "version": "3.2.1"
+  "version": "4.3.0"
 }
