@@ -1,27 +1,27 @@
 {
-  "_from": "boxen@^1.2.1",
-  "_id": "boxen@1.3.0",
+  "_from": "boxen@^4.2.0",
+  "_id": "boxen@4.2.0",
   "_inBundle": false,
-  "_integrity": "sha512-TNPjfTr432qx7yOjQyaXm3dSR0MH9vXp7eT1BFSl/C51g+EFnOR9hTg1IreahGBmDNCehscshe45f+C1TBZbLw==",
+  "_integrity": "sha512-eB4uT9RGzg2odpER62bBwSLvUeGC+WbRjjyyFhGsKnc8wp/m0+hQsMUvUe3H2V0D5vw0nBdO1hCJoZo5mKeuIQ==",
   "_location": "/boxen",
   "_phantomChildren": {},
   "_requested": {
     "type": "range",
     "registry": true,
-    "raw": "boxen@^1.2.1",
+    "raw": "boxen@^4.2.0",
     "name": "boxen",
     "escapedName": "boxen",
-    "rawSpec": "^1.2.1",
+    "rawSpec": "^4.2.0",
     "saveSpec": null,
-    "fetchSpec": "^1.2.1"
+    "fetchSpec": "^4.2.0"
   },
   "_requiredBy": [
     "/update-notifier"
   ],
-  "_resolved": "https://registry.npmjs.org/boxen/-/boxen-1.3.0.tgz",
-  "_shasum": "55c6c39a8ba58d9c61ad22cd877532deb665a20b",
-  "_spec": "boxen@^1.2.1",
-  "_where": "C:\\Users\\ahmad\\Desktop\\2dv513\\node_modules\\update-notifier",
+  "_resolved": "https://registry.npmjs.org/boxen/-/boxen-4.2.0.tgz",
+  "_shasum": "e411b62357d6d6d36587c8ac3d5d974daa070e64",
+  "_spec": "boxen@^4.2.0",
+  "_where": "C:\\Users\\ahmad\\Desktop\\2dv513-Assignment-3-Database-Theory\\backend_nodeJs\\node_modules\\update-notifier",
   "author": {
     "name": "Sindre Sorhus",
     "email": "sindresorhus@gmail.com",
@@ -32,27 +32,31 @@
   },
   "bundleDependencies": false,
   "dependencies": {
-    "ansi-align": "^2.0.0",
-    "camelcase": "^4.0.0",
-    "chalk": "^2.0.1",
-    "cli-boxes": "^1.0.0",
-    "string-width": "^2.0.0",
-    "term-size": "^1.2.0",
-    "widest-line": "^2.0.0"
+    "ansi-align": "^3.0.0",
+    "camelcase": "^5.3.1",
+    "chalk": "^3.0.0",
+    "cli-boxes": "^2.2.0",
+    "string-width": "^4.1.0",
+    "term-size": "^2.1.0",
+    "type-fest": "^0.8.1",
+    "widest-line": "^3.1.0"
   },
   "deprecated": false,
   "description": "Create boxes in the terminal",
   "devDependencies": {
-    "ava": "*",
-    "nyc": "^11.0.3",
-    "xo": "*"
+    "ava": "^2.1.0",
+    "nyc": "^14.1.1",
+    "tsd": "^0.11.0",
+    "xo": "^0.25.3"
   },
   "engines": {
-    "node": ">=4"
+    "node": ">=8"
   },
   "files": [
-    "index.js"
+    "index.js",
+    "index.d.ts"
   ],
+  "funding": "https://github.com/sponsors/sindresorhus",
   "homepage": "https://github.com/sindresorhus/boxen#readme",
   "keywords": [
     "cli",
@@ -73,7 +77,7 @@
     "url": "git+https://github.com/sindresorhus/boxen.git"
   },
   "scripts": {
-    "test": "xo && nyc ava"
+    "test": "xo && nyc ava && tsd"
   },
-  "version": "1.3.0"
+  "version": "4.2.0"
 }
