@@ -1,27 +1,27 @@
 {
-  "_from": "camelcase@^4.0.0",
-  "_id": "camelcase@4.1.0",
+  "_from": "camelcase@^5.3.1",
+  "_id": "camelcase@5.3.1",
   "_inBundle": false,
-  "_integrity": "sha1-1UVjW+HjPFQmScaRc+Xeas+uNN0=",
+  "_integrity": "sha512-L28STB170nwWS63UjtlEOE3dldQApaJXZkOI1uMFfzf3rRuPegHaHesyee+YxQ+W6SvRDQV6UrdOdRiR153wJg==",
   "_location": "/camelcase",
   "_phantomChildren": {},
   "_requested": {
     "type": "range",
     "registry": true,
-    "raw": "camelcase@^4.0.0",
+    "raw": "camelcase@^5.3.1",
     "name": "camelcase",
     "escapedName": "camelcase",
-    "rawSpec": "^4.0.0",
+    "rawSpec": "^5.3.1",
     "saveSpec": null,
-    "fetchSpec": "^4.0.0"
+    "fetchSpec": "^5.3.1"
   },
   "_requiredBy": [
     "/boxen"
   ],
-  "_resolved": "https://registry.npmjs.org/camelcase/-/camelcase-4.1.0.tgz",
-  "_shasum": "d545635be1e33c542649c69173e5de6acfae34dd",
-  "_spec": "camelcase@^4.0.0",
-  "_where": "C:\\Users\\ahmad\\Desktop\\2dv513\\node_modules\\boxen",
+  "_resolved": "https://registry.npmjs.org/camelcase/-/camelcase-5.3.1.tgz",
+  "_shasum": "e3c9b31569e106811df242f715725a1f4c494320",
+  "_spec": "camelcase@^5.3.1",
+  "_where": "C:\\Users\\ahmad\\Desktop\\2dv513-Assignment-3-Database-Theory\\backend_nodeJs\\node_modules\\boxen",
   "author": {
     "name": "Sindre Sorhus",
     "email": "sindresorhus@gmail.com",
@@ -32,16 +32,18 @@
   },
   "bundleDependencies": false,
   "deprecated": false,
-  "description": "Convert a dash/dot/underscore/space separated string to camelCase: foo-bar → fooBar",
+  "description": "Convert a dash/dot/underscore/space separated string to camelCase or PascalCase: `foo-bar` → `fooBar`",
   "devDependencies": {
-    "ava": "*",
-    "xo": "*"
+    "ava": "^1.4.1",
+    "tsd": "^0.7.1",
+    "xo": "^0.24.0"
   },
   "engines": {
-    "node": ">=4"
+    "node": ">=6"
   },
   "files": [
-    "index.js"
+    "index.js",
+    "index.d.ts"
   ],
   "homepage": "https://github.com/sindresorhus/camelcase#readme",
   "keywords": [
@@ -56,7 +58,9 @@
     "separator",
     "string",
     "text",
-    "convert"
+    "convert",
+    "pascalcase",
+    "pascal-case"
   ],
   "license": "MIT",
   "name": "camelcase",
@@ -65,10 +69,7 @@
     "url": "git+https://github.com/sindresorhus/camelcase.git"
   },
   "scripts": {
-    "test": "xo && ava"
+    "test": "xo && ava && tsd"
   },
-  "version": "4.1.0",
-  "xo": {
-    "esnext": true
-  }
+  "version": "5.3.1"
 }
