@@ -1,59 +1,55 @@
 {
-  "_from": "chalk@^2.0.1",
-  "_id": "chalk@2.4.2",
+  "_from": "chalk@^3.0.0",
+  "_id": "chalk@3.0.0",
   "_inBundle": false,
-  "_integrity": "sha512-Mti+f9lpJNcwF4tWV8/OrTTtF1gZi+f8FqlyAdouralcFWFQWF2+NgCHShjkCb+IFBLq9buZwE1xckQU4peSuQ==",
+  "_integrity": "sha512-4D3B6Wf41KOYRFdszmDqMCGq5VV/uMAB273JILmO+3jAlh8X4qDtdtgCR3fxtbLEMzSx22QdhnDcJvu2u1fVwg==",
   "_location": "/chalk",
   "_phantomChildren": {},
   "_requested": {
     "type": "range",
     "registry": true,
-    "raw": "chalk@^2.0.1",
+    "raw": "chalk@^3.0.0",
     "name": "chalk",
     "escapedName": "chalk",
-    "rawSpec": "^2.0.1",
+    "rawSpec": "^3.0.0",
     "saveSpec": null,
-    "fetchSpec": "^2.0.1"
+    "fetchSpec": "^3.0.0"
   },
   "_requiredBy": [
     "/boxen",
     "/update-notifier"
   ],
-  "_resolved": "https://registry.npmjs.org/chalk/-/chalk-2.4.2.tgz",
-  "_shasum": "cd42541677a54333cf541a49108c1432b44c9424",
-  "_spec": "chalk@^2.0.1",
-  "_where": "C:\\Users\\ahmad\\Desktop\\2dv513\\node_modules\\update-notifier",
+  "_resolved": "https://registry.npmjs.org/chalk/-/chalk-3.0.0.tgz",
+  "_shasum": "3f73c2bf526591f574cc492c51e2456349f844e4",
+  "_spec": "chalk@^3.0.0",
+  "_where": "C:\\Users\\ahmad\\Desktop\\2dv513-Assignment-3-Database-Theory\\backend_nodeJs\\node_modules\\update-notifier",
   "bugs": {
     "url": "https://github.com/chalk/chalk/issues"
   },
   "bundleDependencies": false,
   "dependencies": {
-    "ansi-styles": "^3.2.1",
-    "escape-string-regexp": "^1.0.5",
-    "supports-color": "^5.3.0"
+    "ansi-styles": "^4.1.0",
+    "supports-color": "^7.1.0"
   },
   "deprecated": false,
   "description": "Terminal string styling done right",
   "devDependencies": {
-    "ava": "*",
-    "coveralls": "^3.0.0",
-    "execa": "^0.9.0",
-    "flow-bin": "^0.68.0",
-    "import-fresh": "^2.0.0",
+    "ava": "^2.4.0",
+    "coveralls": "^3.0.7",
+    "execa": "^3.2.0",
+    "import-fresh": "^3.1.0",
     "matcha": "^0.7.0",
-    "nyc": "^11.0.2",
-    "resolve-from": "^4.0.0",
-    "typescript": "^2.5.3",
-    "xo": "*"
+    "nyc": "^14.1.1",
+    "resolve-from": "^5.0.0",
+    "tsd": "^0.7.4",
+    "xo": "^0.25.3"
   },
   "engines": {
-    "node": ">=4"
+    "node": ">=8"
   },
   "files": [
-    "index.js",
-    "templates.js",
-    "types/index.d.ts",
-    "index.js.flow"
+    "source",
+    "index.d.ts"
   ],
   "homepage": "https://github.com/chalk/chalk#readme",
   "keywords": [
@@ -80,6 +76,7 @@
     "text"
   ],
   "license": "MIT",
+  "main": "source",
   "name": "chalk",
   "repository": {
     "type": "git",
@@ -87,18 +84,13 @@
   },
   "scripts": {
     "bench": "matcha benchmark.js",
-    "coveralls": "nyc report --reporter=text-lcov | coveralls",
-    "test": "xo && tsc --project types && flow --max-warnings=0 && nyc ava"
+    "test": "xo && nyc ava && tsd"
   },
-  "types": "types/index.d.ts",
-  "version": "2.4.2",
+  "version": "3.0.0",
   "xo": {
-    "envs": [
-      "node",
-      "mocha"
-    ],
-    "ignores": [
-      "test/_flow.js"
-    ]
+    "rules": {
+      "unicorn/prefer-string-slice": "off",
+      "unicorn/prefer-includes": "off"
+    }
   }
 }
