@@ -1,8 +1,8 @@
 {
   "_from": "chokidar@^3.2.2",
-  "_id": "chokidar@3.3.1",
+  "_id": "chokidar@3.5.1",
   "_inBundle": false,
-  "_integrity": "sha512-4QYCEWOcK3OJrxwvyyAOxFuhpvOVCYkr33LPfFNBjAD/w3sEzWsp2BUOkI4l9bHvWioAd0rc6NlHUOEaWkTeqg==",
+  "_integrity": "sha512-9+s+Od+W0VJJzawDma/gvBNQqkTiqYTWLuZoyAsivsI4AaWTCzHG06/TMjsf1cYe9Cb97UCEhjz7HvnPk2p/tw==",
   "_location": "/chokidar",
   "_phantomChildren": {},
   "_requested": {
@@ -18,10 +18,10 @@
   "_requiredBy": [
     "/nodemon"
   ],
-  "_resolved": "https://registry.npmjs.org/chokidar/-/chokidar-3.3.1.tgz",
-  "_shasum": "c84e5b3d18d9a4d77558fef466b1bf16bbeb3450",
+  "_resolved": "https://registry.npmjs.org/chokidar/-/chokidar-3.5.1.tgz",
+  "_shasum": "ee9ce7bbebd2b79f49f304799d5468e31e14e68a",
   "_spec": "chokidar@^3.2.2",
-  "_where": "C:\\Users\\ahmad\\Desktop\\2dv513\\node_modules\\nodemon",
+  "_where": "C:\\Users\\ahmad\\Desktop\\2dv513-Assignment-3-Database-Theory\\backend_nodeJs\\node_modules\\nodemon",
   "author": {
     "name": "Paul Miller",
     "url": "https://paulmillr.com"
@@ -42,79 +42,29 @@
   "dependencies": {
     "anymatch": "~3.1.1",
     "braces": "~3.0.2",
-    "fsevents": "~2.1.2",
+    "fsevents": "~2.3.1",
     "glob-parent": "~5.1.0",
     "is-binary-path": "~2.1.0",
     "is-glob": "~4.0.1",
     "normalize-path": "~3.0.0",
-    "readdirp": "~3.3.0"
+    "readdirp": "~3.5.0"
   },
   "deprecated": false,
-  "description": "A neat wrapper around node.js fs.watch / fs.watchFile / fsevents.",
+  "description": "Minimal and efficient cross-platform file watching library",
   "devDependencies": {
-    "@types/node": "^12",
+    "@types/node": "^14",
     "chai": "^4.2",
-    "dtslint": "^2.0.0",
-    "eslint": "^6.6.0",
-    "mocha": "^6.2.2",
-    "nyc": "^14.1.1",
+    "dtslint": "^3.3.0",
+    "eslint": "^7.0.0",
+    "mocha": "^7.0.0",
+    "nyc": "^15.0.0",
     "rimraf": "^3.0.0",
-    "sinon": "^7.5.0",
+    "sinon": "^9.0.1",
     "sinon-chai": "^3.3.0",
     "upath": "^1.2.0"
   },
   "engines": {
     "node": ">= 8.10.0"
-  },
-  "eslintConfig": {
-    "extends": "eslint:recommended",
-    "parserOptions": {
-      "ecmaVersion": 9,
-      "sourceType": "script"
-    },
-    "env": {
-      "node": true,
-      "es6": true
-    },
-    "rules": {
-      "array-callback-return": "error",
-      "no-empty": [
-        "error",
-        {
-          "allowEmptyCatch": true
-        }
-      ],
-      "no-lonely-if": "error",
-      "no-var": "error",
-      "object-shorthand": "error",
-      "prefer-arrow-callback": [
-        "error",
-        {
-          "allowNamedFunctions": true
-        }
-      ],
-      "prefer-const": [
-        "error",
-        {
-          "ignoreReadBeforeAssign": true
-        }
-      ],
-      "prefer-destructuring": [
-        "error",
-        {
-          "object": true,
-          "array": false
-        }
-      ],
-      "prefer-spread": "error",
-      "prefer-template": "error",
-      "radix": "error",
-      "strict": "error",
-      "quotes": [
-        "error",
-        "single"
-      ]
-    }
   },
   "files": [
     "index.js",
@@ -145,7 +95,7 @@
     ]
   },
   "optionalDependencies": {
-    "fsevents": "~2.1.2"
+    "fsevents": "~2.3.1"
   },
   "repository": {
     "type": "git",
@@ -158,5 +108,5 @@
     "test": "npm run lint && npm run mocha"
   },
   "types": "./types/index.d.ts",
-  "version": "3.3.1"
+  "version": "3.5.1"
 }
