@@ -1,31 +1,31 @@
 {
-  "_from": "cli-boxes@^1.0.0",
-  "_id": "cli-boxes@1.0.0",
+  "_from": "cli-boxes@^2.2.0",
+  "_id": "cli-boxes@2.2.1",
   "_inBundle": false,
-  "_integrity": "sha1-T6kXw+WclKAEzWH47lCdplFocUM=",
+  "_integrity": "sha512-y4coMcylgSCdVinjiDBuR8PCC2bLjyGTwEmPb9NHR/QaNU6EUOXcTY/s6VjGMD6ENSEaeQYHCY0GNGS5jfMwPw==",
   "_location": "/cli-boxes",
   "_phantomChildren": {},
   "_requested": {
     "type": "range",
     "registry": true,
-    "raw": "cli-boxes@^1.0.0",
+    "raw": "cli-boxes@^2.2.0",
     "name": "cli-boxes",
     "escapedName": "cli-boxes",
-    "rawSpec": "^1.0.0",
+    "rawSpec": "^2.2.0",
     "saveSpec": null,
-    "fetchSpec": "^1.0.0"
+    "fetchSpec": "^2.2.0"
   },
   "_requiredBy": [
     "/boxen"
   ],
-  "_resolved": "https://registry.npmjs.org/cli-boxes/-/cli-boxes-1.0.0.tgz",
-  "_shasum": "4fa917c3e59c94a004cd61f8ee509da651687143",
-  "_spec": "cli-boxes@^1.0.0",
-  "_where": "C:\\Users\\ahmad\\Desktop\\2dv513\\node_modules\\boxen",
+  "_resolved": "https://registry.npmjs.org/cli-boxes/-/cli-boxes-2.2.1.tgz",
+  "_shasum": "ddd5035d25094fce220e9cab40a45840a440318f",
+  "_spec": "cli-boxes@^2.2.0",
+  "_where": "C:\\Users\\ahmad\\Desktop\\2dv513-Assignment-3-Database-Theory\\backend_nodeJs\\node_modules\\boxen",
   "author": {
     "name": "Sindre Sorhus",
     "email": "sindresorhus@gmail.com",
-    "url": "sindresorhus.com"
+    "url": "https://sindresorhus.com"
   },
   "bugs": {
     "url": "https://github.com/sindresorhus/cli-boxes/issues"
@@ -34,16 +34,19 @@
   "deprecated": false,
   "description": "Boxes for use in the terminal",
   "devDependencies": {
-    "ava": "*",
-    "xo": "*"
+    "ava": "^1.4.1",
+    "tsd": "^0.7.2",
+    "xo": "^0.24.0"
   },
   "engines": {
-    "node": ">=0.10.0"
+    "node": ">=6"
   },
   "files": [
     "index.js",
+    "index.d.ts",
     "boxes.json"
   ],
+  "funding": "https://github.com/sponsors/sindresorhus",
   "homepage": "https://github.com/sindresorhus/cli-boxes#readme",
   "keywords": [
     "cli",
@@ -65,7 +68,7 @@
     "url": "git+https://github.com/sindresorhus/cli-boxes.git"
   },
   "scripts": {
-    "test": "xo && ava"
+    "test": "xo && ava && tsd"
   },
-  "version": "1.0.0"
+  "version": "2.2.1"
 }
