@@ -1,27 +1,27 @@
 {
-  "_from": "configstore@^3.0.0",
-  "_id": "configstore@3.1.2",
+  "_from": "configstore@^5.0.1",
+  "_id": "configstore@5.0.1",
   "_inBundle": false,
-  "_integrity": "sha512-vtv5HtGjcYUgFrXc6Kx747B83MRRVS5R1VTEQoXvuP+kMI+if6uywV0nDGoiydJRy4yk7h9od5Og0kxx4zUXmw==",
+  "_integrity": "sha512-aMKprgk5YhBNyH25hj8wGt2+D52Sw1DRRIzqBwLp2Ya9mFmY8KPvvtvmna8SxVR9JMZ4kzMD68N22vlaRpkeFA==",
   "_location": "/configstore",
   "_phantomChildren": {},
   "_requested": {
     "type": "range",
     "registry": true,
-    "raw": "configstore@^3.0.0",
+    "raw": "configstore@^5.0.1",
     "name": "configstore",
     "escapedName": "configstore",
-    "rawSpec": "^3.0.0",
+    "rawSpec": "^5.0.1",
     "saveSpec": null,
-    "fetchSpec": "^3.0.0"
+    "fetchSpec": "^5.0.1"
   },
   "_requiredBy": [
     "/update-notifier"
   ],
-  "_resolved": "https://registry.npmjs.org/configstore/-/configstore-3.1.2.tgz",
-  "_shasum": "c6f25defaeef26df12dd33414b001fe81a543f8f",
-  "_spec": "configstore@^3.0.0",
-  "_where": "C:\\Users\\ahmad\\Desktop\\2dv513\\node_modules\\update-notifier",
+  "_resolved": "https://registry.npmjs.org/configstore/-/configstore-5.0.1.tgz",
+  "_shasum": "d365021b5df4b98cdd187d6a3b0e3f6a7cc5ed96",
+  "_spec": "configstore@^5.0.1",
+  "_where": "C:\\Users\\ahmad\\Desktop\\2dv513-Assignment-3-Database-Theory\\backend_nodeJs\\node_modules\\update-notifier",
   "author": {
     "name": "Sindre Sorhus",
     "email": "sindresorhus@gmail.com",
@@ -32,21 +32,21 @@
   },
   "bundleDependencies": false,
   "dependencies": {
-    "dot-prop": "^4.1.0",
+    "dot-prop": "^5.2.0",
     "graceful-fs": "^4.1.2",
-    "make-dir": "^1.0.0",
-    "unique-string": "^1.0.0",
-    "write-file-atomic": "^2.0.0",
-    "xdg-basedir": "^3.0.0"
+    "make-dir": "^3.0.0",
+    "unique-string": "^2.0.0",
+    "write-file-atomic": "^3.0.0",
+    "xdg-basedir": "^4.0.0"
   },
   "deprecated": false,
   "description": "Easily load and save config without having to think about where and how",
   "devDependencies": {
-    "ava": "*",
-    "xo": "*"
+    "ava": "^2.1.0",
+    "xo": "^0.24.0"
   },
   "engines": {
-    "node": ">=4"
+    "node": ">=8"
   },
   "files": [
     "index.js"
@@ -56,7 +56,6 @@
     "config",
     "store",
     "storage",
-    "conf",
     "configuration",
     "settings",
     "preferences",
@@ -75,5 +74,5 @@
   "scripts": {
     "test": "xo && ava"
   },
-  "version": "3.1.2"
+  "version": "5.0.1"
 }
