@@ -1,27 +1,27 @@
 {
-  "_from": "crypto-random-string@^1.0.0",
-  "_id": "crypto-random-string@1.0.0",
+  "_from": "crypto-random-string@^2.0.0",
+  "_id": "crypto-random-string@2.0.0",
   "_inBundle": false,
-  "_integrity": "sha1-ojD2T1aDEOFJgAmUB5DsmVRbyn4=",
+  "_integrity": "sha512-v1plID3y9r/lPhviJ1wrXpLeyUIGAZ2SHNYTEapm7/8A9nLPoyvVp3RK/EPFqn5kEznyWgYZNsRtYYIWbuG8KA==",
   "_location": "/crypto-random-string",
   "_phantomChildren": {},
   "_requested": {
     "type": "range",
     "registry": true,
-    "raw": "crypto-random-string@^1.0.0",
+    "raw": "crypto-random-string@^2.0.0",
     "name": "crypto-random-string",
     "escapedName": "crypto-random-string",
-    "rawSpec": "^1.0.0",
+    "rawSpec": "^2.0.0",
     "saveSpec": null,
-    "fetchSpec": "^1.0.0"
+    "fetchSpec": "^2.0.0"
   },
   "_requiredBy": [
     "/unique-string"
   ],
-  "_resolved": "https://registry.npmjs.org/crypto-random-string/-/crypto-random-string-1.0.0.tgz",
-  "_shasum": "a230f64f568310e1498009940790ec99545bca7e",
-  "_spec": "crypto-random-string@^1.0.0",
-  "_where": "C:\\Users\\ahmad\\Desktop\\2dv513\\node_modules\\unique-string",
+  "_resolved": "https://registry.npmjs.org/crypto-random-string/-/crypto-random-string-2.0.0.tgz",
+  "_shasum": "ef2a7a966ec11083388369baa02ebead229b30d5",
+  "_spec": "crypto-random-string@^2.0.0",
+  "_where": "C:\\Users\\ahmad\\Desktop\\2dv513-Assignment-3-Database-Theory\\backend_nodeJs\\node_modules\\unique-string",
   "author": {
     "name": "Sindre Sorhus",
     "email": "sindresorhus@gmail.com",
@@ -34,21 +34,21 @@
   "deprecated": false,
   "description": "Generate a cryptographically strong random string",
   "devDependencies": {
-    "ava": "*",
-    "xo": "*"
+    "ava": "^1.4.1",
+    "tsd": "^0.7.2",
+    "xo": "^0.24.0"
   },
   "engines": {
-    "node": ">=4"
+    "node": ">=8"
   },
   "files": [
-    "index.js"
+    "index.js",
+    "index.d.ts"
   ],
   "homepage": "https://github.com/sindresorhus/crypto-random-string#readme",
   "keywords": [
     "random",
     "string",
-    "str",
-    "rand",
     "text",
     "id",
     "identifier",
@@ -66,10 +66,7 @@
     "url": "git+https://github.com/sindresorhus/crypto-random-string.git"
   },
   "scripts": {
-    "test": "xo && ava"
+    "test": "xo && ava && tsd"
   },
-  "version": "1.0.0",
-  "xo": {
-    "esnext": true
-  }
+  "version": "2.0.0"
 }
